Wrote migrations, haven't run "php artisan migrate " yet

+++ database/migrations/2024_06_07_210730_create_reviews_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('reviews', function (Blueprint $table) {
             $table->id();
             $table->timestamps();
-            $table->integer('user_id');
-            $table->integer('hotel_id');
+            $table->bigInteger('user_id');
+            $table->bigInteger('hotel_id');
             $table->integer('rating');
             $table->text('comment');
             $table->foreign('user_id')->references('id')->on('users');
